requirements.txt
beautifulsoup4==4.9.3
discord.py==1.5.1
requests==2.25.0
psycopg2-binary